Location modules.

+++ LIKE/LIKE/LKExternal/POP/POPSpringAnimationInternal.h
@@ -46,7 +46,7 @@ struct _POPSpringAnimationState : _POPPropertyAnimationState
       const CGFloat *previous2Values = previous2Vec->data();
 
       for (NSUInteger idx = 0; idx < count; idx++) {
-        if ((fabsf(toValues[idx] - previousValues[idx]) >= t) || (fabsf(previous2Values[idx] - previousValues[idx]) >= t)) {
+        if ((fabsf((float)toValues[idx] - (float)previousValues[idx]) >= (float)t) || (fabsf((float)previous2Values[idx] - (float)previousValues[idx]) >= (float)t)) {
           return false;
         }
       }
